Rename isIdentity to isFixedPoint
There's an example in lecture 5 typeclasses incarnations in the chapter Categorical FP:

```scala
def isIdentity[A: Eq](a: A)(f: A => A): Boolean = {
  implicitly[Eq[A]].eqv(a, f(a))
}
```

The function name `isIdentity` suggests that it returns wether `f` is the identity function. But I think it only checks wether `a` is a fixed point of `f`. So I would propose to rename it.